requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-crontab==0.7.1
psycopg2-binary==2.9.9

